Using factor(1.5) to set max.map.tasks at runtime
The default value of mapred.tasktracker.map.tasks.maximum is 2 which
means that at max only 2 maps can run per tasktracker. With variance
in cluster instances, this value can't be static.

For the time being, the recipe deals with it by using the floor of
total number of cores times the factor(1.5).

Hence, max.map.tasks is 1 for m1.medium, 3 for m1.large and 6 for
m1.xlarge.

Signed-off-by: Shashank Sahni <shredder12@gmail.com>